Move it into `dialog`
